I had written a code
Vending machines are a great way to earn money. If you are able to place a vending machine in a location that has plenty of traffic, the money will start flowing in and all you need to do is to program the vending machines so that the machine knows what items are stocked and how much each item costs. You will need to order inventory and restock the machine at planned intervals.
The User get menu similar to real vending machine and from there any drink can be ordered after paying money in dollar it also shows left over drinks