Updatet mos_algorithm.cpp
Mos algorithm is a general technique and is supposed to be a template. Depending on the problem, we have to provide the above three functions, and it will handle everything. 
Although I have updated the file to contain an example.
Problem: Given an array of n elements, and q queries. Each query contains a range and we have to find the sum of elements in the range. Although it can be easily solved by prefix sums, here I have tried to solve it via the Mos algorithm, which can also help in solving some other hard problems.