Add files via upload
/*Tower of Hanoi */

/*
This is a mathematical puzzle where we have three rods and n disks. The objective of the puzzle is to move the entire stack to another rod, obeying 2 rules: 
(i) Only one disk can be moved at a time. (ii) No bigger disk may be placed on top of a smaller disk.
*/