Golmine Solution
Chef and Chefu are working as gold miners. There are a total of NN gold mines, numbered 11 through NN. For each valid ii, the ii-th gold mine contains GiGi gold in total; if only Chef worked in it, it would take him AiAi days to completely mine it, while if only Chefu worked in it, it would take him BiBi days.

Each of our miners may only work in one mine at a time, but they may decide to start working in another mine at any time (even in the middle of some day), any number of times. They also choose the mines to work in independently from each other and they may work in the same mine at the same time. Mining gold is a continuous process, i.e. if a miner works for tt days (where tt is a real number) in a mine where this miner has mining speed gg gold per day, then he will mine g⋅tg⋅t gold. Obviously, it is impossible to work in a mine after no gold remains in it. For example, if a gold mine contains 3030 gold and Chef needs 22 days to completely mine it, but he spends 11 day in it, then he will mine 1515 gold; if Chefu needs 11 day to completely mine the same gold mine, and both Chef and Chefu start working in this mine at the same time, it will be empty after 2/32/3 days ― Chefu will mine 2020 gold, while Chef will mine 1010 gold.

At each point of time, both Chef and Chefu know the gold mine in which the other miner is working. Each of them wants to gather the maximum amount of gold for himself. Find the amounts of gold the miners will have if they both act optimally.